restructure redux code for search functionality
