+++ client/redux/actions/index.js
@@ -1,2 +1,3 @@
 export * from './noteActions'
+export * from './searchActions'
 export * from './userActions'
