+++ client/redux/actions/noteActions.js
@@ -136,27 +136,6 @@ export const checkItem = (params) => (dispatch) => {
   });
 }
 
-// Sets the term that the user is searching for
-export const setSearchTerm = (term) => (dispatch) => {
-  return new Promise((resolve, reject) => {
-    resolve(dispatch(setSearchTermAction(term)));
-  });
-}
-
-// Sets the term that the user is searching for (params are name, item, type)
-export const addSearchFilter = (params) => (dispatch) => {
-  return new Promise((resolve, reject) => {
-    resolve(dispatch(addSearchFilterAction(params)));
-  });
-}
-
-// Sets the term that the user is searching for (params is name)
-export const removeSearchFilter = (params) => (dispatch) => {
-  return new Promise((resolve, reject) => {
-    resolve(dispatch(removeSearchFilterAction(params)));
-  });
-}
-
 // ACTION CREATORS
 // equivalent to " ... => { return { ... } } "
 
@@ -260,19 +239,4 @@ const switchBodyAction = (note) => ({
   meta: {
     emit: true
   }
-});
-
-const setSearchTermAction = (term) => ({
-  type: noteConstants.SET_SEARCH_TERM,
-  payload: term
-});
-
-const addSearchFilterAction = (params) => ({
-  type: noteConstants.ADD_SEARCH_FILTER,
-  payload: params
-});
-
-const removeSearchFilterAction = (params) => ({
-  type: noteConstants.REMOVE_SEARCH_FILTER,
-  payload: params
 });
