+++ client/redux/constants/index.js
@@ -1,2 +1,3 @@
 export * from './noteConstants'
+export * from './searchConstants'
 export * from './userConstants'
