+++ client/redux/constants/noteConstants.js
@@ -18,9 +18,5 @@ export const noteConstants = {
   ADD_ITEM: 'ADD_ITEM',
   UPDATE_ITEM: 'UPDATE_ITEM',
   REMOVE_ITEM: 'REMOVE_ITEM',
-  CHECK_ITEM: 'CHECK_ITEM',
-
-  SET_SEARCH_TERM: 'SET_SEARCH_TERM',
-  ADD_SEARCH_FILTER: 'ADD_SEARCH_FILTER',
-  REMOVE_SEARCH_FILTER: 'REMOVE_SEARCH_FILTER'
+  CHECK_ITEM: 'CHECK_ITEM'
 }
