+++ client/redux/constants/searchConstants.js
@@ -1,0 +1,6 @@
+// Constants used for Search Reducer and Search Actions
+export const searchConstants = {
+  SET_SEARCH_TERM: 'SET_SEARCH_TERM',
+  ADD_SEARCH_FILTER: 'ADD_SEARCH_FILTER',
+  REMOVE_SEARCH_FILTER: 'REMOVE_SEARCH_FILTER'
+}
